+++ clusterutil/cluster_validation_config_test.go
@@ -1,13 +1,31 @@
 package clusterutil
 
 import (
+	"flag"
 	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 )
 
-func TestClusterValidationProtocolConfigValidate(t *testing.T) {
+func TestClusterValidationConfig_RegisteredFlags(t *testing.T) {
+	cfg := ClusterValidationConfig{}
+	// Before we track registered flags, no registered flags is returned.
+	require.Empty(t, cfg.RegisteredFlags())
+
+	fs := flag.NewFlagSet("test", flag.PanicOnError)
+	cfg.RegisterFlagsWithPrefix("prefix", fs)
+
+	// After we track registered flags, the label flag is returned.
+	registeredFlags := cfg.RegisteredFlags()
+	require.NotEmpty(t, registeredFlags)
+	require.Equal(t, "prefix", registeredFlags.Prefix)
+	require.Len(t, registeredFlags.Flags, 1)
+	_, ok := registeredFlags.Flags["label"]
+	require.True(t, ok)
+}
+
+func TestClusterValidationProtocolConfig_Validate(t *testing.T) {
 	testCases := map[string]struct {
 		label          string
 		enabled        bool
@@ -59,4 +77,25 @@ func TestClusterValidationProtocolConfigValidate(t *testing.T) {
 			require.Equal(t, testCase.expectedErr, err)
 		})
 	}
+}
+
+func TestServerClusterValidationConfig_RegisteredFlags(t *testing.T) {
+	var cfg = ServerClusterValidationConfig{}
+	// Before we track registered flags, no registered flags is returned.
+	require.Empty(t, cfg.RegisteredFlags())
+
+	fs := flag.NewFlagSet("test", flag.PanicOnError)
+	cfg.RegisterFlagsWithPrefix("server.cluster-validation.", fs)
+
+	// After we track registered flags, label, grpc.enabled and grpc.soft-validation flags are returned.
+	registeredFlags := cfg.RegisteredFlags()
+	require.NotEmpty(t, registeredFlags)
+	require.Equal(t, "server.cluster-validation.", registeredFlags.Prefix)
+	require.Len(t, registeredFlags.Flags, 3)
+	expectedFlags := []string{"label", "grpc.enabled", "grpc.soft-validation"}
+	for _, flagName := range expectedFlags {
+		_, ok := registeredFlags.Flags[flagName]
+		require.True(t, ok)
+	}
+
 }